src/Domain/Client.java
package Domain;

import java.util.Objects;

public class Client extends Entity{
    private String name, firstName, CNP, dateOfBirth, dateOfRegistration;

    public Client(int id, String name, String firstName, String CNP, String dateOfBirth, String dateOfRegistration) {
        super(id);
        this.name = name;
        this.firstName = firstName;
        this.CNP = CNP;
        this.dateOfBirth = dateOfBirth;
        this.dateOfRegistration = dateOfRegistration;
    }

    @Override
    public String toString() {
        return name + " " + firstName + " " + CNP + " " + dateOfBirth + " " + dateOfRegistration;
    }

    @Override
    public int hashCode() {
        return Objects.hash(super.hashCode(), getName(), getFirstName(), getCNP(), getDateOfBirth(), getDateOfRegistration());
    }

    public String getDateOfRegistration() {
        return dateOfRegistration;
    }

    public void setDateOfRegistration(String dateOfRegistration) {
        this.dateOfRegistration = dateOfRegistration;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getFirstName() {
        return firstName;
    }

    public void setFirstName(String firstName) {
        this.firstName = firstName;
    }

    public String getCNP() {
        return CNP;
    }

    public void setCNP(String CNP) {
        this.CNP = CNP;
    }

    public String getDateOfBirth() {
        return dateOfBirth;
    }

    public void setDateOfBirth(String dateOfBirth) {
        this.dateOfBirth = dateOfBirth;
    }
}